feat: add empty state
refs: CLX-2515 , CLX-2518
builds: Student
affects: Student
release note: none
