+++ Horizon/Horizon/Sources/Features/Learn/Learn/ProgramSwitcher/ProgramSwitcherCourseView.swift
